Finish with ExcHandler

+++ src/main/java/Basic/HW/service/CarServiceImpl.java
@@ -3,6 +3,7 @@ package Basic.HW.service;
 import Basic.HW.dto.Car;
 import Basic.HW.dto.Driver;
 import Basic.HW.dto.request.CarRequest;
+import Basic.HW.exception.ExcHandler.TypicalError;
 import Basic.HW.repository.CarRepo;
 import Basic.HW.repository.DriverRepo;
 import Basic.HW.dto.response.CarResponse;
@@ -23,11 +24,11 @@ public class CarServiceImpl implements CarService {
      final DriverRepo driverRepo;
 
     @Override
-    public CarResponse saveCar(CarRequest car) {
+    public CarResponse saveCar(CarRequest car) throws ServiceException {
         Driver driver=driverRepo.findByUsername(car.getDriverName());
         if (driver==null){
             log.error("Driver {} not exist for add car",  car.getDriverName());
-            throw new RuntimeException("Driver "+car.getDriverName()+" not exist for add car");
+            throw new ServiceException("Driver "+car.getDriverName()+" not exist for add car", TypicalError.NOT_FOUND);
         }
         log.info("Save new car ");
         Car car1=carRepo.save(new Car(null, car.getType(), car.getColor(),driver));
@@ -35,11 +36,11 @@ public class CarServiceImpl implements CarService {
     }
 
     @Override
-    public void addCarToDriver(Long idCar, String nameDriver)  {
+    public void addCarToDriver(Long idCar, String nameDriver) throws ServiceException {
         Driver driver =driverRepo.findByUsername(nameDriver);
         if(driver==null){
             log.error("Driver {} not exist for add car",  nameDriver);
-            throw new RuntimeException("Driver "+nameDriver+" not exist for add car");
+            throw new ServiceException("Driver "+nameDriver+" not exist for add car",TypicalError.NOT_FOUND);
         }
         carRepo.getById(idCar).setDriver(driver);
         log.info("add new car id={} to Driver {}", idCar,nameDriver);
@@ -47,9 +48,13 @@ public class CarServiceImpl implements CarService {
     }
 
     @Override
-    public CarResponse getCar(Long id) {
-       CarResponse carResponse= convertToCarResponse(carRepo.getById(id));
+    public CarResponse getCar(Long id) throws ServiceException {
         log.info("Fetching  car id={}",  id);
+       CarResponse carResponse= convertToCarResponse(carRepo.getById(id));
+        if(carResponse==null){
+            log.error("Car id={} not exist ",  id);
+            throw new ServiceException("Car id={} "+id+" not exist ",TypicalError.NOT_FOUND);
+        }
         return carResponse;
     }
 
@@ -65,9 +70,11 @@ public class CarServiceImpl implements CarService {
     }
 
    @Override
-    public void deleteCar(Long id) {
-
-
+    public void deleteCar(Long id) throws ServiceException {
+       if(carRepo.getById(id)==null){
+           log.error("Car id={} not exist ",  id);
+           throw new ServiceException("Car id={} "+id+" not exist ",TypicalError.NOT_FOUND);
+       }
         carRepo.delete(carRepo.getById(id));
        log.info("Delete car id={}",id);
