+++ src/main/java/Basic/HW/service/DriverServiceImpl.java
@@ -4,6 +4,7 @@ import Basic.HW.dto.Car;
 import Basic.HW.dto.Driver;
 import Basic.HW.dto.Role;
 import Basic.HW.dto.request.DriverRequest;
+import Basic.HW.exception.ExcHandler.TypicalError;
 import Basic.HW.repository.CarRepo;
 import Basic.HW.repository.DriverRepo;
 import Basic.HW.repository.RoleRepo;
@@ -28,9 +29,9 @@ public class DriverServiceImpl implements DriverService {
 
 
     @Override
-    public Driver saveDriver(DriverRequest driver) {
+    public Driver saveDriver(DriverRequest driver) throws ServiceException {
         if (driverRepo.findByUsername(driver.getUsername()) != null) {
-            throw new RuntimeException("Driver exist");
+            throw new ServiceException("Driver exist",TypicalError.DUPLICATE);
         }
         log.info("save new Driver {} to DB", driver.getUsername());
         driverRepo.save(new Driver(driver.getUsername(), driver.getPassword()));
@@ -48,11 +49,11 @@ public class DriverServiceImpl implements DriverService {
     }
 
     @Override
-    public Driver addRoleToDriver(String nameDriver, String nameRole) {
+    public Driver addRoleToDriver(String nameDriver, String nameRole) throws ServiceException {
         Driver driver = driverRepo.findByUsername(nameDriver);
         Role role = roleRepo.findByName(nameRole);
         if (driver.getRoles().contains(role)) {
-            throw new RuntimeException("Driver has this role");
+            throw new ServiceException("Driver has this role",TypicalError.DUPLICATE);
         }
         driver.getRoles().add(role);
         log.info("add new Role {} to Driver {}", nameRole, nameDriver);
@@ -62,12 +63,12 @@ public class DriverServiceImpl implements DriverService {
     }
 
     @Override
-    public DriverResponse getDriver(String name) {
+    public DriverResponse getDriver(String name) throws ServiceException {
         log.info("Fetching  Driver {}", name);
         Driver driver = driverRepo.findByUsername(name);
         if (driver == null) {
             log.error("Driver {} not exist ", name);
-            throw new RuntimeException("Driver " + name + " not exist ");
+            throw new ServiceException("Driver " + name + " not exist ", TypicalError.NOT_FOUND);
         }
         DriverResponse driverResponse = converterToDriverResponse(driver);
         return driverResponse;
@@ -85,8 +86,12 @@ public class DriverServiceImpl implements DriverService {
     }
 
     @Override
-    public void deleteDriver(String name) {
+    public void deleteDriver(String name) throws ServiceException {
         Driver driver = driverRepo.findByUsername(name);
+        if (driver == null) {
+            log.error("Driver {} not exist ", name);
+            throw new ServiceException("Driver " + name + " not exist ", TypicalError.NOT_FOUND);
+        }
         for (Car car :
                 carRepo.getByDriver_Id(driver.getId())) {
             carRepo.delete(carRepo.getById(car.getId()));
